cleanup(core): consolidate napi bindings
