+++ packages/nx/src/native/metrics/mod.rs
@@ -1,3 +1,2 @@
 pub mod collector;
-pub mod napi_bindings;
 pub mod types;
